+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/enum/enum.tsx
@@ -13,11 +13,11 @@
  *
  **/
 import * as React from 'react'
-import MarionetteRegionContainer from '../marionette-region-container'
-import withListenTo, { WithBackboneProps } from '../backbone-container'
+import MarionetteRegionContainer from '../container/marionette-region-container'
+import withListenTo, { WithBackboneProps } from '../container/backbone-container'
 import { hot } from 'react-hot-loader'
-const PropertyView = require('../../../component/property/property.view.js')
-const PropertyModel = require('../../../component/property/property.js')
+const PropertyView = require('../../component/property/property.view.js')
+const PropertyModel = require('../../component/property/property.js')
 
 type Option = {
   label: string
