+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/map-settings/presentation.tsx
@@ -15,7 +15,7 @@
 import * as React from 'react'
 import styled from 'styled-components'
 import { hot } from 'react-hot-loader'
-import Enum from '../container/enum'
+import Enum from '../enum'
 import ExampleCoordinates from './example-coordinates'
 
 type Props = {
